requirements.txt
# Core dependencies
pyyaml>=6.0
feedparser>=6.0
requests>=2.31
beautifulsoup4>=4.12

# Google AI
google-generativeai>=0.8

# YouTube API
google-api-python-client>=2.100
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.2

# Substack (unofficial)
python-substack>=1.0

# Testing
pytest>=8.0
pytest-cov>=5.0

# Linting
ruff>=0.5
